create custom_head with the favicon inclusion code (#32)
Earlier, the proposal was to modify the deafult.html head section to add the favicon link. However, after discussions in a-chacon/awesome-jekyll-theme#32, it looks like we can accomplish the same thing by adding custom_head.html to the _includes directory without have to change the main theme files. So this PR takes that route.